[TASK] Show image of distribution even if not installed
Previously the image of a distribution extension
was only shown in the extensionmanager, in case
the distribution was already installed. In the
list view, all distributions therefore displayed
the same fallback image.

This is now fixed by adding a new field to the
extension record, which is filled via the XML
parser. As it contains the URL to the image,
it can be directly used in the corresponding
templates and therefore allows to remove the
DistributionImageViewHelper. Instead, a new
web component handles the image and ensures
a preview image is displayed, even if the
distribution image is not available.

Resolves: #83465
Releases: master
Change-Id: Id249b99833571024e39ed3f5991751e24e2e8d1d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69099
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Thomas Löffler <loeffler@spooner-web.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>