[BUGFIX] Ensure throwStatus in Extbase stops the execution workflow
Extbase's throwStatus() method is mentioned that
the execution workflow indeed should be stopped
right away, but never did, as all of TSFE's code
was executed again and again.

Since TYPO3 v11 (#93007) we now have a "PropagateResponseException"
and native PSR-7 Responses in Extbase, which allows us
to do exactly that: Stop and have the Application emit
the response directly with the status code.

Resolves: #94131
Releases: master
Change-Id: I3a6d2785603abcce4399378d0dabc2a5d05d6017
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69135
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>