[BUGFIX] Fix Drag+Drop in Page Module in Workspaces
When using Drag+Drop in Page Module a different
functionality is used than using Clipboard (Move+Paste)
while moving a record in the page module.

This results in a different behaviour when dealing in Workspaces.

The patch uses the same "cmd" instead of "data"
for DataHandler now for Drag+Drop than in Clipboard.

Resolves: #92849
Releases: master, 10.4
Change-Id: Idb8566b166a18c514d149e53699489b7b9eb9247
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69126
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>