[TASK] increase z-index of extbase-debugger
increase z-index for extbase-debugger to from 999 to 99990 (same as the
admin-panel) to prevent the debug output to be hidden behind elements
with z-index values >1000 which is quite common, f.e. happens with
the introduction package.
Side note: this impacts the debug viewhelper output as it uses the
DebuggerUtility::var_dump.

Resolves: #93357
Releases: master
Change-Id: I77e0696dfd19b74d2e8e5d76795965b443fe65ea
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69153
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>