[TASK] Avoid usages of sys_language in site configuration
This patch removes the last remains of sys_language in the
site configurations' TCA.

Therefore, the TCA of `site` and `site_langauge` is changed
to always retrieve possible site languages via an itemsProcFunc,
instead of using a relation to sys_language.

As new site languages now have to be created in the module
directly, a new internal TCA type "siteLanguage" is introduced.
The new type behaves similar to type "inline", but contains some
necessary features, e.g. unique record selector box next to a
"create new" button, which are not available in type "inline".
Also some not needed functionality is omitted.

Instead of the sys_language records, all available site languages
from all existing site configurations are now presented in the
selector box. On selecting one of them, a new site language record
is created, with most of the fields pre filled.

Resolves: #94399
Releases: master
Change-Id: I60ac5b4259aa3c9d90a4aba9881bc1dc2341b464
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69188
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>