[BUGFIX] Allow editing and copying of composer proposal
The composer manifest proposal, introduced in #93931,
uses t3editor to display the composer.json content.
The corresponding TCA was configured to be "readOnly",
which however had no effect until #94081.

Since there is no reason to disallow editing or
copying of the proposal, the "readOnly" option
is removed from the TCA configuration.

Resolves: #93931
Releases: master, 10.4
Change-Id: Ia0c0377d69a21bd599875079fe93c22435d44b4e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69193
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>