[BUGFIX] Allow copying content from readonly t3editor field
With #94081, a new option was introduced to
set a FormEngine t3editor field to "readOnly",
as it was already possible for default text fields.

Using the option "nocursor" however also disabled
copying of the content, which is not intended and
also not in line with the behaviour of "readOnly" for
normal text elements.

Therefore, when "readOnly" is set in TCA for renderType
t3editor, the codeMirror related option is set to TRUE,
which still renders the field as readOnly, but allows to
copy its content again.

For further reference see:

https://codemirror.net/doc/manual.html#option_readOnly

Resolves: #94169
Related: #94081
Releases: master
Change-Id: I33a08c009c22de194bd757a39e7043f3f6b2523e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69212
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>