[BUGFIX] Check request object in ExtensionComposerStatus
Although the ExtensionComposerStatus report implements the
RequestAwareStatusProviderInterface, there is no guarantee
that a PSR-7 request is passed to getStatus().

To prevent type errors we have to ensure, the request
object is set, before accessing its values.

This patch therefore checks the request object before
adding the returnUrl argument to the query parameters.

Resolves: #94160
Releases: master, 10.4
Change-Id: Iba3481e80a86d72e6520a5eb4dd427dec3fa4207
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69215
Tested-by: André Buchmann <andy.schliesser@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: André Buchmann <andy.schliesser@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>