[BUGFIX] Remove legacy foreign_selector conversion
During TYPO3 v7 the FormEngine got rewritten, and a "todo"
was left, which can be removed as it now contains wrong
information as this information has been processed already
in a prior FormEngine DataProvider.

It is removed because the array implode implodes a
array of arrays building $record[$foreign_selector] (e.g. FAL)
into $record[$foreign_selector] = 'Array'; which was never
used within the InlineReordContainer anymore.

Resolves: #94177
Releases: master
Change-Id: I0d31c640271f6aeb45477557deba3985789e81bf
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69218
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>