[BUGFIX] Define missing fieldChangeFunc property
The composer proposal view is using FormEngine
to render the composer.json content with t3editor.

Therefore, a "fake" TCA is prepared. In case, EXT:t3editor
is not installed, FormEngine automatically falls back to
TextElement. Unlike t3editor, this element requires
the property `fieldChangeFunc` to be set and will
otherwise trigger a PHP warning while calling implode()
on it.

The PHP warning is fixed by defining the property in
the fake TCA.

As you may noticed, it's not always the best idea to use
FormEngine for non-form views. A follow-up patch will
therefore replace the FormEngine rendering with direct
usage of our codemirror web component, which however
is master only.

Resolves: #94179
Releases: master, 10.4
Change-Id: Iab1c7485c5fd365c5e643434c690f8d19e143d20
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69223
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>