[TASK] Deprecate Application->run callback
The run() method for all Applications contained
a callback which was added in TYPO3 v7, before
proper solutions for subrequests and a more
streamlined boot process, was in place.

The additional code can now be handled via
PSR-15 middlewares. Therefore, the callback
argument in Application->run() is now marked
as deprecated.

Resolves: #94272
Releases: master
Change-Id: I9fe00e7b3663553dadc1b48b89c3083537206c5e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69258
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>