[BUGFIX] Pass PSR-7 Request to RequestBuilder->build()
Since #93853, RequestBuilder->build() expects
a PSR-7 Request to be passed.

Because the FLUIDTEMPLATE content object was
not adjusted accordingly, an exception is triggered
as soon as extbase variables are defined for the
content object.

This is fixed by passing the current PSR-7 Request
to the method. The PSR-7 Request is available in any
content object since #92984.

Resolves: #94203
Related: #93853
Releases: master
Change-Id: Ife3b9a076757e4af8b384e49ab00c945411d47c6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69293
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>