[BUGFIX] Update modal instances stack when hiding a modal
The modal stack handling is slightly changed to update the instance stack
when a modal was requested to get closed already and not once a modal has
been destroyed anymore.

This fixes an issue in case a closing modal triggers a new modal which
resulted in the new modal being broken.

Resolves: #94219
Releases: master, 10.4
Change-Id: Idbc16ac08e95d8b3fed896783672b58ff062cafe
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69331
Tested-by: core-ci <typo3@b13.com>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>