[FEATURE] Simplify sharing of backend urls
Atfer a lot of preperation in #93048 and #93988,
it's finally possible to share URLs to TYPO3 backend
modules. Even special modules such as FormEngine.

To ease the use for editors, the ShortcutButton
is extended for a new option. If enabled, which
is the default behaviour, the shortcut button
in the module header is replaced. The new button
allows to open a dropdown with the additional
possibility to copy the URL of the current
page to the operating systems' clipboard. Next
to the already exisiting "Add shortcut" option.

Since those URLs should not contain the user
specific token, the UriBuilder features a new
constant "SHAREABLE_URL". If set as $referenceType
in one of the supporting methods, e.g. buildUriFromRoute(),
an absoulte URL without the token is returned.

To copy the URL to the operating systems' clipboard,
a new web component "CopyToClipboard" is introduced.
This component is added without any dependency to
the URL sharing functionality and can therefore be
freely used for other backend components as well.

For the new button, the font awesome "share-alt"
icon is registered in the IconRegistry.

Resolves: #93921
Releases: master
Change-Id: Id1dcfe1f2af764fbe000659ddb49a7369322d5b6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69338
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>