[TASK] Remove a library dependent view helper test
The removed unit test mocks the system under test
and checks for specific method calls. The test
relies on AbstractTagBasedViewHelper from
TYPO3Fluid\Fluid which added a feature in 2.7.0.
This fails in core v9 and v10 which allow 2.6
and 2.7. This whitebox test is kinda useless
anyway, so we drop it.

Resolves: #94249
Releases: master, 10.4, 9.5
Change-Id: I70804f2881a84352ca793e7d523531a583a44b34
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69357
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>