[TASK] Log symfony deprecation messages
Symfony deprecation messages are prefixed by an @-sign in order to
only be thrown if an explicit error handler for E_USER_DEPRECATED
is registered.

Other than for regular silenced errors, we want these deprecation
messages to be logged in any case, and therefore ignore
the @-sign for E_USER_DEPRECATED errors now.

Releases: master
Resolves: #94269
Change-Id: I2d1fb443ef0bf5d01361f1980ab74b94a116fb1d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69366
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>