[BUGFIX] Keep preview info on forwarding in EditDocumentController
The "View" respectively "Save & View" functionality of
EditDocumentController is as following:

1. Clicking the corresponding button saves the form and
a JavaScript event handler opens a new blank window

2. After the data was persisted, the form reloads and
the preview code (using the ImmediateAction component)
is written to the module body. This update the previously
opened window with the correct preview URL.

However, since #94114, a redirect is performed as soon
as the data has been persisted. This is done to ensure
persisting is only done for POST requests. Since the
information whether the form was stored using the "View"
respectively "Save & View" is not forwarded in the redirect,
the new window remains blank.

This is now fixed by adding the corresponding information
to the redirect URL. Also a now unnecessary code block in
processRequest() is removed.

Resolves: #94287
Releases: master
Change-Id: I4f9be3fb286a6cdfe6e1d3e071eb0c6df0801f5e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69415
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>