[TASK] Add link to "Show all versions" view for installed extensions
Besides all available versions of a TER extension,
the "Show all versions" view in EXT:extensionmanager
also displays general information such as the category
or the dependencies, making it more or less an extension
detail view.

Since those information are also relevant for already
installed extensions, it's now possible to access this
view in the "Installed extensions" list, using a new
link on the extensions' version string.

Resolves: #94298
Releases: master
Change-Id: Iab131e5494f143315abd607de1a79f5f0571a2dc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69429
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>