[BUGFIX] Fix terms in Info > Page TSconfig
In the select list in "Info > Page TSconfig" the technical
representation of the terms (as they are used when setting
TSconfig) is now shown in brackets, e.g. [RTE.].

This is consistent with showing the technical representation
in other places in the backend, such as the page id in the
page layout or page tree, the table fields and Flexform
fields in the edit forms etc.

Where there are descriptive terms, (such as "Module: Web>Page")
a combination of the descriptive terms and the technical
terms is used, such as "Module: Web>Page [mod.web_layout.]"

The explanation "with overruling user settings" for mod is
omitted. This would make the list more cluttered and visually
distracting. If it is necessary to point this out, it
might be added as a note in the results.

Resolves: #94323
Related: #94322
Releases: master, 10.4
Change-Id: I728f3d6399f040356d7bee4e30c3f33baf9703d4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69467
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>