[TASK] Deprecate extbase setDispatched() and isDispatched()
Extbase Request setDispatched() and isDispatched()
violate the PSR-7 'immutable' concept and need to fall.

With the changes from #92502 where extbase actions
usually return a ResponseInterface, together with the
deprecation of StopActionException in #94351, we don't
need this information in the Request anymore: It can
be handled internally within the dispatcher, depending
on the type of the response.

The patch deprecates the two methods in extbase request
and adapts usages towards an internal flag. We
throw a deprecation warning since the flag is needed for
b/w compat in v11, though.

Resolves: #94394
Related: #94351
Related: #92502
Releases: master
Change-Id: I1c70daa7f5b0c55e17e4b7c815d260ca68cef23c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69470
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>