[BUGFIX] Store preview of FAL Online Media in public folder
When using Online Media (Youtube, Vimeo) with FAL and with
composer-mode, the preview images are put under var/transient/
which isn't accessible to the outside.

This change adapts the change to be put in typo3temp/assets/online_media/ to be always accessible.

In the future, this API should be refactored to have preview images
based on processed files instead.

Resolves: #93860
Related: #84545
Releases: master, 10.4, 9.5
Change-Id: I086b962a0efad2bcc0c67679838d3d973812462d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70213
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>