[BUGFIX] Use correct slide direction in MultiStepWizard
Bootstrap beta3 introduced a regression, which reversed
the slide direction. This therefore broke our implementation,
since we are using an event listener, evaluating the direction.

Because we were not aware of this being "just" a regression
instead of an intended API change, we switched the direction
in our event listener in #94037, too.

In the final 5.0 release, we updated to in #94089, the regression
has been fixed, breaking our event listener once again.

This is fixed by restoring the initial and correct direction check.

Related bootstrap PRs:

- https://github.com/twbs/bootstrap/pull/32913
- https://github.com/twbs/bootstrap/pull/33499

Resolves: #94338
Releases: master
Change-Id: I0bc4dc2dfd37935fc43e04f311bd850916c31004
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69486
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>