[BUGFIX] Avoid crash due to endless loop in Fluid-based Page Module
Page Module partial `PageLayout/Record/Preview` tries to render optional
`PageLayout/Record/{item.record.CType}/Preview` and falls back to a raw
representation if that custom partial does not exist.

In case `CType` is empty, this resolves to `PageLayout/Record//Preview`
which is actually using `PageLayout/Record/Preview.html` again - that's
the partial again that initiated the process - which will after some
seconds crash due to an endless loop of that partial.

In order to make things explicit, the defaults are now located in
their dedicated namespace `RecordDefault` to avoid ambiguity.

* `PageLayout/Record/Preview` -> `PageLayout/RecordDefault/Preview`
* `PageLayout/Record/Header` -> `PageLayout/RecordDefault/Header`
* `PageLayout/Record/Footer` -> `PageLayout/RecordDefault/Footer`

Resolves: #94343
Releases: 10.4, master
Change-Id: I0c0385a9d89561de2395fc81124e471ee9fed8dd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69493
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>