[BUGFIX] Respect offline storages on context menu initialization
The FileProvider tries to retrieve the file / folder object
on initialization. In case the requested folder or storage
is temporarily offline, a ResourceDoesNotExistException
is thrown.

This is now fixed by catching the exception an setting
the $record to NULL. This allows to still render the
context menu with the remaining options, such as edit
the storage record or display the records information.

Resolves: #94372
Releases: master, 10.4
Change-Id: I2ef6437b0d6e22462d73c9536433d7dc1914f680
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69517
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>