[BUGFIX] Undefined key warning in table wizard for PHP8
Open the table wizard with empty test in the text area
or a click on the save button on table wizard page
thow undefined key warning.
This will avoid php warnings by adding a fallback
value or a check if the key is present.

Added acceptance tests for the table wizard to  make
sure the wizard will work as expected.

Resolves: #94417
Releases: master
Change-Id: I9283c9203f71710c8bffcc74432781d2159dcde7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69575
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>