[TASK] Remove all code occurrences from showRemovedLocalizationRecords
The TCA option showRemovedLocalizationRecords was former used to enable
showing 'dangling' records (localized records without parent). Because
it was subject to a broken db structure, its usage was removed in patch
f6f36d3. But there was still code mentioning this key without using it.

A TCA migration removes this option and logs a message which pinpoints
the lack of usage.

Resolves: #94441
Releases: master
Change-Id: I8c32bb845512585a6314d92f8fcafffd959279ad
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69665
Tested-by: core-ci <typo3@b13.com>
Tested-by: Simon Gilli <typo3@gilbertsoft.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>