[TASK] Allow to distinguish between file and folder rows in FileList
To allow further improvements of the FileList, it is
important to be able to distinguish between file and
folder rows.

This is done by adding data attributes to the corresponding
table rows. Besides the type (file or folder), also relevant
identifiers, e.g. folder identifier or sys_file uid, are added.

Additionally the addElement() method, building the table
rows, is cleaned up as it was obviously just copied from
DatabaseRecordList and therefore contained a lot of unused
and - in this context - wrong functionality (e.g. fetching a
possible l10n_parent from the data array).

Resolves: #94522
Releases: master
Change-Id: Ib4dbccb8d0957956ca46cdacdefb7e1a71fb9fe7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69799
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>