[BUGFIX] Fix several issues regarding Recordlist search
Since #93892, clearing the search field did not longer
submit the form. This is now fixed by adding a dedicated
JS module, listening on the browsers "search" event.

When using the search in the Record selector, the search
options dropdown is now not longer cut off, in case no
search results are present.

Additionally, the functionality to toggle the search field
in the recordlist module is now moved into the Recordlist
JS module, as it does not belong to EXT:backend and
furthermore does not require a dedicated JS module.

Resolves: #94463
Resolves: #94557
Releases: master
Change-Id: I504a27fc6cb3d3689555169ac3e39813e2029544
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69830
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Guido Schmechel <guido.schmechel@brandung.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>