[FEATURE] Allow overriding fileFolder config with TSconfig
The "fileFolder" configuration options, available for TCA
columns of type "select" are used to fill the select field
with predefined files (images / icons). Nowadays this is
frequently used to make a corporate icon set available for
editors. In multi site installations however, those icon sets
usually differ from site to site.

Therefore, the AbstractItemProvider is now extended to allow
overriding those settings with TSconfig (TCEFORM).

Furthermore, to streamline the TCA configuration and to be
in line with the corresponding overrides, the "fileFolder"
TCA configuration options are moved into a dedicated sub array
"fileFolderConfiguration" and the properties are renamed
to be consistent with other TCA options.

* fileFolder => folder
* fileFolder_extList => allowedExtensions
* fileFolder_recursions => depth

A TCA migration is in place.

Resolves: #94406
Releases: master
Change-Id: I621198523edfd328ad68d692d9194017c445406f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69832
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>