[FEATURE] Introduce TCA type "category"
A new TCA type "category" is introduced, which
allows to simplify the configuration of category
TCA columns. Besides the benefit for integrators,
this allows to deprecate the CategoryRegistry
in the next step.

The new type can also be used for other use cases.
Therefore, the TCA option "relationship" is available
for this TCA type. Besides "manyToMany" (default), this
can also be set to "oneToOne" or "oneToMany".

Using the new type, FormEngine will always render a
category tree. This means, no additional `renderType`
is defined. In such case, TCA type "select" can
still be used as before, without any limitation. However,
all relevant places in core are adjusted in this patch.

The category element is rendered through a custom element
(web component), reducing inline javascript.

Resolves: #94622
Releases: master
Change-Id: I1b95c42288b070fa6bac114266f5ad246a045b21
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69899
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>