[BUGFIX] Check whether tt_content parent page is accessible
Both, DatabaseRecordList as well as the ContextMenu feature the
"Show" button to preview a content element on its parent page.
However, in case the parent page is a "no view doktype" (e.g. sys
folder), those buttons lead to a 404 error.

This is now fixed by properly checking whether a content elements'
parent page can be viewed. If not, the button is no longer shown.

Resolves: #93718
Releases: master, 10.4
Change-Id: I2ad48ee7e44d06f569496c4bed2bbd172791b86c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69959
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>