[BUGFIX] Clarify IPv6 address matching
When matching IPv6 addresses (for example in IPmaskList)
appending the address with * as wildcards will not work.
The function GeneralUtility::cmpIPv6() uses validIPv6()
which will fail on an address with *.

Using addresses with CIDR-notation, will work for both
IPv4 and IPv6 addresses (e.g. 2001:0DB8:1234::/48,
192.16.0.0/16).

The description of TYPO3_CONF_VARS and comments for the
function in GeneralUtility is updated accordingly.

Resolves: #94620
Releases: master, 10.4
Change-Id: Ie068ca579a41cd3c6552e6a302bbe11b0ca2883a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70161
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>